+++ android/ui/src/main/java/com/trafi/ui/components/OutlinedTextField.kt
@@ -1,4 +1,4 @@
-package com.trafi.ui
+package com.trafi.ui.components
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
